Introduce TodosModel, which AddForm and TodoList talk to.
Their view models manipulate and read (respectively) from the TodosModel. Thus there's no more publishing of "add" events, but instead direct manipulation of/binding to the model. Excellent!
