+++ lib/start.js
@@ -20,12 +20,12 @@ var TodosPageTodoList = require("../components/TodosPage/TodoList");
 var AboutPage = require("../components/AboutPage");
 
 var Storage = require("./Storage");
-var createBindingList = require("./todos").createBindingList;
+var TodosModel = require("./TodosModel");
 
 function buildTodosPage(todos, nav) {
     var appBar = new TodosPageAppBar();
     var navBar = new TodosPageNavBar(nav);
-    var addForm = new TodosPageAddForm();
+    var addForm = new TodosPageAddForm(todos);
     var todoList = new TodosPageTodoList(todos, appBar.setCommandContext);
 
     return new TodosPage(appBar, navBar, addForm, todoList);
@@ -37,7 +37,7 @@ function buildAboutPage(nav) {
 
 function buildApp() {
     var storage = new Storage(window.localStorage);
-    var todos = createBindingList(storage.loadTodos());
+    var todos = new TodosModel(storage.loadTodos());
 
     var todosPage = buildTodosPage(todos, nav);
     var aboutPage = buildAboutPage(nav);
